Improve ui, add goBack button, add filtering pipe, reuse recipe-list
