+++ src/app/app.module.ts
@@ -4,18 +4,21 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { AppNavbarComponent } from './components/app-navbar/app-navbar.component';
-import { AppCategoriesComponent } from './components/app-categories/app-categories.component';
-import { AppRecipeComponent } from './components/app-recipe/app-recipe.component';
-import { AppFavoritesComponent } from './components/app-favorites/app-favorites.component';
-import { AppCategoryRecipesComponent } from './components/app-category-recipes/app-category-recipes.component';
-import { AppNewRecipeComponent } from './components/app-new-recipe/app-new-recipe.component';
+import { AppNavbarComponent } from './components/navbar/navbar.component';
+import { AppCategoriesComponent } from './components/categories/categories.component';
+import { AppRecipeComponent } from './components/recipe/recipe.component';
+import { AppFavoritesComponent } from './components/favorites/favorites.component';
+import { AppCategoryRecipesComponent } from './components/category-recipes/category-recipes.component';
+import { AppNewRecipeComponent } from './components/new-recipe/new-recipe.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { StoreModule, StoreRootModule } from '@ngrx/store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { reducers } from './store/reducers';
 import { EffectsModule } from '@ngrx/effects';
 import { RecipeCategoriesEffect } from './store/effects';
+import { FilterRecordsPipe } from './pipes/filter-records.pipe';
+import { LoadMoreButtonComponent } from './shared/load-more-button/load-more-button.component';
+import { RecipeListComponent } from './shared/recipe-list/recipe-list.component';
 
 @NgModule({
   declarations: [
@@ -25,7 +28,10 @@ import { RecipeCategoriesEffect } from './store/effects';
     AppRecipeComponent,
     AppFavoritesComponent,
     AppCategoryRecipesComponent,
-    AppNewRecipeComponent
+    AppNewRecipeComponent,
+    FilterRecordsPipe,
+    LoadMoreButtonComponent,
+    RecipeListComponent
   ],
   imports: [
     BrowserModule,
