--- src/app/components/app-categories/app-categories.component.html
@@ -1,16 +1,0 @@
-<app-navbar></app-navbar>
-<div
-  class="categories"
-  *ngIf="!(isLoading$ | async)"
->
-<div
-  class="category"
-  *ngFor="let category of recipeCategories$ | async"
-  [routerLink]="['/category', { categoryName: category.name }]"
-  [skipLocationChange]="true"
->
-  <p class="category-name" >{{ category.name }}</p>
-  <img [src]="category.image" />
-  <p class="category-description">{{ category.description }}</p>
-</div>
-</div>
