--- src/app/components/app-categories/app-categories.component.scss
@@ -1,39 +1,0 @@
-.categories {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    flex-direction: row;
-    gap: 20px;
-}
-
-.category {
-    width: 20%;
-    border: 2px solid blue;
-    padding: 5px;
-    border-color: blue;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-
-    img {
-        max-width: 200px;
-        max-height: 200px
-    }
-
-    .category-name {
-        line-height: 2;
-        font-size: 20px;
-        font-weight: 700;
-    }
-
-    .category-description {
-        overflow: hidden;
-        text-overflow: ellipsis;
-        display: -webkit-box;
-        -webkit-line-clamp: 3;
-        -webkit-box-orient: vertical;
-        line-height: 1.5;
-    }
-}
-
