--- src/app/components/app-category-recipes/app-category-recipes.component.scss
@@ -1,29 +1,0 @@
-.category-recipes {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    flex-direction: row;
-    gap: 20px;
-}
-
-.recipe {
-    width: 20%;
-    border: 2px solid blue;
-    padding: 5px;
-    border-color: blue;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-
-    img {
-        max-width: 200px;
-        max-height: 200px
-    }
-
-    .recipe-name {
-        line-height: 2;
-        font-size: 20px;
-        font-weight: 700;
-    }
-}
