--- src/app/components/app-navbar/app-navbar.component.html
@@ -1,6 +1,0 @@
-<div class="navbar">
-    <p class="title">{{ pageTitle }}</p>
-    <a routerLink="/favorites" [skipLocationChange]="true">Go to favorites</a>
-    <a routerLink="/categories" [skipLocationChange]="true">Go to categories</a>
-    <a routerLink="/new-recipe" [skipLocationChange]="true">Create new recipe</a>
-</div>
