--- src/app/components/app-navbar/app-navbar.component.scss
@@ -1,16 +1,0 @@
-.navbar {
-    height: 50px;
-    width: 100%;
-    background-color: aqua;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 10px;
-    padding-left: 20px;
-    padding-right: 20px;
-}
-
-p {
-    font-size: 20px;
-    font-weight: 700;
-}
