--- src/app/components/app-navbar/app-navbar.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AppNavbarComponent } from './app-navbar.component';
-
-describe('AppNavbarComponent', () => {
-  let component: AppNavbarComponent;
-  let fixture: ComponentFixture<AppNavbarComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [AppNavbarComponent]
-    });
-    fixture = TestBed.createComponent(AppNavbarComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
