--- src/app/components/app-new-recipe/app-new-recipe.component.html
@@ -1,32 +1,0 @@
-<app-navbar></app-navbar>
-<form class="recipe" [formGroup]="recipeForm" (ngSubmit)="handleSubmit()">
-  <div class="recipe-name">
-    <label for="name">Name: </label>
-    <input type="text" id="name" formControlName="name" />
-  </div>
-  <div class="recipe-category">
-    <label for="category">Category: </label>
-    <select formControlName="category">
-        <option 
-            *ngFor="let category of recipeDataSource.recipeCategories"
-            [value]="category.name"
-        >{{category.name}}
-    </option>
-    </select>
-  </div>
-  <div class="recipe-area">
-    <label for="area">Area: </label>
-    <select formControlName="area">
-        <option 
-            *ngFor="let area of recipeDataSource.recipeAreas"
-            [value]="area"
-        >{{area}}
-    </option>
-    </select>
-  </div>
-  <div class="recipe-instructions">
-    <label for="instructions">Instructions: </label>
-    <input type="text" id="instructions" formControlName="instructions" />
-  </div>
-  <button type="submit" [disabled]="recipeForm.invalid">Submit</button>
-</form>
