--- src/app/components/app-new-recipe/app-new-recipe.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AppNewRecipeComponent } from './app-new-recipe.component';
-
-describe('AppNewRecipeComponent', () => {
-  let component: AppNewRecipeComponent;
-  let fixture: ComponentFixture<AppNewRecipeComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [AppNewRecipeComponent]
-    });
-    fixture = TestBed.createComponent(AppNewRecipeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
