+++ src/app/components/categories/categories.component.html
@@ -1,0 +1,26 @@
+<app-navbar></app-navbar>
+
+<div *ngIf="!(isLoading$ | async) && (recipeCategories$ | async) as categories">
+  <div class="categories">
+    <div
+      class="category"
+      *ngFor="let category of categories | filterRecords : recordsLimit"
+      [routerLink]="['/category', { categoryName: category.name }]"
+      [skipLocationChange]="true"
+    >
+      <div class="category-content">
+        <div class="category-image">
+          <img [src]="category.image" alt="{{ category.name }} Image" />
+        </div>
+        <div class="category-details">
+          <p class="category-name">{{ category.name }}</p>
+          <p class="category-description">{{ category.description }}</p>
+        </div>
+      </div>
+    </div>
+  </div>
+  <app-load-more-button
+    *ngIf="recordsLimit < categories.length"
+    (onClick)="handleAddMore()"
+  ></app-load-more-button>
+</div>
