+++ src/app/components/categories/categories.component.scss
@@ -1,0 +1,60 @@
+.categories {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  gap: 20px;
+  padding: 20px;
+}
+
+.category {
+  width: 23%;
+  background-color: #fff;
+  border: 2px solid #3498db;
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease;
+
+  &:hover {
+    transform: scale(1.05);
+  }
+
+  .category-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .category-image {
+    overflow: hidden;
+    border-bottom: 2px solid #3498db;
+
+    img {
+      max-width: 100%;
+      height: auto;
+      border-radius: 10px 10px 0 0;
+    }
+  }
+
+  .category-details {
+    padding: 15px;
+
+    .category-name {
+      line-height: 1.5;
+      font-size: 18px;
+      font-weight: 700;
+      color: #333;
+      margin-bottom: 8px;
+    }
+
+    .category-description {
+      overflow: hidden;
+      text-overflow: ellipsis;
+      display: -webkit-box;
+      -webkit-line-clamp: 3;
+      -webkit-box-orient: vertical;
+      line-height: 1.5;
+      color: #666;
+    }
+  }
+}
