+++ src/app/components/favorites/favorites.component.html
@@ -1,0 +1,6 @@
+<app-navbar></app-navbar>
+
+<app-recipe-list
+  [recipes]="recipes"
+  [recipeLimit]="'8'"
+></app-recipe-list>
