+++ src/app/components/favorites/favorites.component.scss
