+++ src/app/components/new-recipe/new-recipe.component.html
@@ -1,0 +1,36 @@
+<app-navbar></app-navbar>
+
+<form class="recipe" [formGroup]="recipeForm" (ngSubmit)="handleSubmit()">
+  <div class="form-group">
+    <label for="name">Name:</label>
+    <input type="text" id="name" formControlName="name" />
+  </div>
+
+  <div class="form-group">
+    <label for="category">Category:</label>
+    <select formControlName="category">
+      <option
+        *ngFor="let category of recipeDataSource.recipeCategories"
+        [value]="category.name"
+      >
+        {{ category.name }}
+      </option>
+    </select>
+  </div>
+
+  <div class="form-group">
+    <label for="area">Area:</label>
+    <select formControlName="area">
+      <option *ngFor="let area of recipeDataSource.recipeAreas" [value]="area">
+        {{ area }}
+      </option>
+    </select>
+  </div> 
+
+  <div class="form-group">
+    <label for="instructions">Instructions:</label>
+    <textarea id="instructions" formControlName="instructions"></textarea>
+  </div>
+
+  <button type="submit" [disabled]="recipeForm.invalid">Submit</button>
+</form>
