+++ src/app/components/recipe/recipe.component.scss
@@ -1,0 +1,70 @@
+.recipe {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+  background-color: #f9f9f9;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+
+  .recipe-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+
+    .recipe-name {
+      font-size: 24px;
+      font-weight: 700;
+    }
+
+    .favorite-button {
+      padding: 10px;
+      font-size: 14px;
+      cursor: pointer;
+      background-color: #3498db;
+      color: #fff;
+      border: none;
+      border-radius: 5px;
+
+      &:hover {
+        background-color: #2980b9;
+      }
+    }
+  }
+
+  .recipe-image {
+    max-width: 100%;
+    height: auto;
+    border-radius: 8px;
+  }
+
+  .recipe-details {
+    text-align: center;
+
+    p {
+      margin: 0;
+    }
+
+    .recipe-category,
+    .recipe-area {
+      font-size: 16px;
+      margin-bottom: 10px;
+    }
+  }
+
+  .recipe-instructions {
+    max-width: 80%;
+
+    h2 {
+      font-size: 20px;
+      margin-bottom: 10px;
+    }
+
+    p {
+      font-size: 16px;
+      line-height: 1.6;
+    }
+  }
+}
