+++ src/app/pipes/filter-records.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { FilterRecordsPipe } from './filter-records.pipe';
+
+describe('FilterRecordsPipe', () => {
+  it('create an instance', () => {
+    const pipe = new FilterRecordsPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
