+++ src/app/shared/load-more-button/load-more-button.component.html
@@ -1,0 +1,5 @@
+<div class="load-more">
+    <button
+      (click)="handleLoadMore()"
+    >Load more</button>
+</div>
