+++ src/app/shared/load-more-button/load-more-button.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { LoadMoreButtonComponent } from './load-more-button.component';
+
+describe('LoadMoreButtonComponent', () => {
+  let component: LoadMoreButtonComponent;
+  let fixture: ComponentFixture<LoadMoreButtonComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [LoadMoreButtonComponent]
+    });
+    fixture = TestBed.createComponent(LoadMoreButtonComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
