+++ src/app/shared/recipe-list/recipe-list.component.scss
@@ -1,0 +1,47 @@
+.category-recipes {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  gap: 20px;
+  padding: 20px;
+}
+
+.recipe {
+  width: 23%;
+  background-color: #fff;
+  border: 2px solid #3498db;
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease;
+
+  &:hover {
+    transform: scale(1.05);
+  }
+
+  .recipe-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .recipe-image {
+    overflow: hidden;
+    border-bottom: 2px solid #3498db;
+
+    img {
+      max-width: 100%;
+      height: auto;
+      border-radius: 10px 10px 0 0;
+    }
+  }
+
+  .recipe-name {
+    padding: 15px;
+    line-height: 1.5;
+    font-size: 18px;
+    font-weight: 700;
+    color: #333;
+    text-align: center;
+  }
+}
